add new assments

+++ assiment/ass36.java
@@ -3,9 +3,9 @@
 class  Test{
   public static void main(String args[]){
     int diameter = 15;
-    int radius = diameter/2;
-    int height = 7;
-    float surface_area =   2*(3.14)*radius*height;
+    float  radius = (float)diameter/2;
+    float  height = 7;
+    float surface_area =   2*(3.14f)*radius*height;
     System.out.println(surface_area);
   }
 }
